requirements.txt
discord.py
docker
mechanize